# Clone instead of dup record when readonlyifying fetched records